login/login_app/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
boto3==1.26.110
Pillow==10.0.0
PyJWT==2.6.0
qrcode==7.3.1
